test removing int32

+++ train_flowers.py
@@ -79,8 +79,8 @@ DATA_SET = 'Flowers'
 
 
 def load_data():
-    x = np.load(DATA_SET + '/flower_imgs.npy').astype(np.int32)
-    y = np.load(DATA_SET + '/flower_labels.npy').astype(np.int32)
+    x = np.load(DATA_SET + '/flower_imgs.npy')
+    y = np.load(DATA_SET + '/flower_labels.npy')
     channeled = np.reshape(np.rollaxis(x, 3) / 255., (3, -1))
     return train_test_split(x, y, test_size=0.15), channeled.mean(1), channeled.std(1)
